website/requirements.txt
asgiref==3.8.1
beautifulsoup4==4.12.3
certifi==2024.8.30
charset-normalizer==3.3.2
Django==4.2.16
django-bootstrap-v5==1.0.11
django-environ==0.11.2
django-extensions==3.2.3
django-redis==5.4.0
djangorestframework==3.15.2
idna==3.8
MarkupSafe==2.1.5
pytz==2024.1
redis==5.0.8
requests==2.32.3
setuptools==74.1.1
soupsieve==2.6
sqlparse==0.5.1
urllib3==2.2.2
Werkzeug==3.0.4
zope.interface==7.0.3
